fid-metrics/requirements.txt
hydra-core
opencv-python
pillow
rich
scipy
torch
torchvision
gradio